Add MIT License and acknowledgments
Added the MIT License to the project to clearly define the terms under which the software can be used, modified, and distributed. This license allows for both commercial and non-commercial use with the condition that the copyright notice and permission notice are included in all copies or substantial portions of the Software. 

Additionally, included acknowledgments for the contributions from Aarush Raj Gandhari, V.R. Sai Sharanya, and other team members who supported the development of the project.